Replace SQLite withjj Postgres (#53)
* Replace sqlite gem for pg

* Update database config file

* Update database migrations

* Update ESI planetary colony logic

* Use rails credentials for ESI

* Add region and constellation seeds

+++ db/migrate/20201211043127_create_items.rb
@@ -1,7 +1,7 @@
 class CreateItems < ActiveRecord::Migration[6.0]
   def change
     create_table :items, id: false do |t|
-      t.integer :id, null: false
+      t.bigint :id, null: false
       t.string :name
 
       t.index :id, unique: true
